update web browser to latest version.

+++ libs/NetworkService/xiaomi_migration_manager.cpp
@@ -1,6 +1,8 @@
 #include "xiaomi_migration_manager.h"
+#include "xiaomi_account_manager.h"
 #include "access_manager.h"
 #include "cookie_jar.h"
+#include "ns_utils.h"
 
 #include "Device/DeviceInfo.h"
 #include "Device/fat.h"
@@ -19,6 +21,7 @@ static const QString MIGRATION_START_URI = "/dk_id/migrate/kindle/start";
 static const QString MIGRATION_OK_URI = "/dk_id/migrate/kindle/success";
 static const QString MIGRATION_FAIL_URI = "/dk_id/migrate/kindle/fail";
 static const QString MIGRATION_CANCEL_URI = "/dk_id/migrate/kindle/cancel";
+static const QString MI_DUSHU_URI = "http://login.dushu.xiaomi.com";
 
 static const QString MIGRATION_COOKIE_CHECKPOINT = "/dk_id/migrate";
 static const qreal MIGRATION_ZOOM = 1.0;
@@ -49,11 +52,15 @@ QString MigrationServerConfiguration::migrationServerCancel()
 
 XiaomiMigrationManager::XiaomiMigrationManager(const QString& host)
     : view_(0)
+    , account_manager_(new XiaomiAccountManager)
     , config_(host)
 {
     connect(this, SIGNAL(migrationSucceeded()), SLOT(onMigrationSucceeded()));
     connect(this, SIGNAL(migrationCanceled()), SLOT(onMigrationCanceled()));
     connect(this, SIGNAL(migrationFailed()), SLOT(onMigrationFailed()));
+
+    connect(account_manager_.get(), SIGNAL(pageChanged(const QString&)), SIGNAL(pageChanged(const QString&)));
+    connect(account_manager_.get(), SIGNAL(loginFinished(bool)), this, SLOT(onXiaomiExchangeFinished(bool)));
 }
 
 XiaomiMigrationManager::~XiaomiMigrationManager()
@@ -116,9 +123,15 @@ qreal XiaomiMigrationManager::getZoomFactor()
     return MIGRATION_ZOOM;
 }
 
-bool XiaomiMigrationManager::start()
+bool XiaomiMigrationManager::start(const QString& duokan_account_id)
 {
+    if (duokan_account_id.isEmpty())
+    {
+        return false;
+    }
+    duokan_account_id_ = duokan_account_id;
     XiaomiMigration::removeFile();
+
     QString start_uri = config_.migrationServerStart();
     QUrl url = guessUrlFromString(start_uri);
     addCookiesForEntry(url);
@@ -135,13 +148,25 @@ void XiaomiMigrationManager::addCookiesForEntry(const QUrl& url)
     QByteArray app_id_name("app_id");
     QByteArray kindle_user("KindleUser");
     QNetworkCookie app_id_cookie(app_id_name, kindle_user);
+    app_id_cookie.setDomain(".duokan.com");
+    app_id_cookie.setPath("/");
     
     QByteArray device_id_name("device_id");
     QByteArray device_id(DeviceInfo::GetDeviceID().c_str());
     QNetworkCookie device_id_cookie(device_id_name, device_id);
+    device_id_cookie.setDomain(".duokan.com");
+    device_id_cookie.setPath("/");
+    qDebug("Add Device ID to Cookie:%s", device_id.constData());
+
+    QByteArray user_id_name("user_id");
+    QNetworkCookie user_id_cookie(user_id_name, duokan_account_id_.toLocal8Bit());
+    user_id_cookie.setDomain(".duokan.com");
+    user_id_cookie.setPath("/");
+    qDebug("Add Account ID to Cookie:%s", qPrintable(duokan_account_id_));
     
     cookies.push_back(app_id_cookie);
     cookies.push_back(device_id_cookie);
+    cookies.push_back(user_id_cookie);
     xiaomi_cookie_jar->setCookiesFromUrl(cookies, url);
 }
     
@@ -152,17 +177,23 @@ bool XiaomiMigrationManager::loadDataFromCookies(const QUrl& url)
     
     foreach(QNetworkCookie cookie, cookies)
     {
-        if (cookie.name() == "mi_token")
+        if (cookie.name() == "token")
         {
             migration_result_.mutableToken() = QString::fromLocal8Bit(cookie.value().data());
         }
-        else if (cookie.name() == "mi_device_id")
+        else if (cookie.name() == "device_id")
         {
             migration_result_.mutableDeviceId() = QString::fromLocal8Bit(cookie.value().data());
         }
-        else if (cookie.name() == "mi_user_id")
+        else if (cookie.name() == "user_id")
         {
-            migration_result_.mutableMiId() = QString::fromLocal8Bit(cookie.value().data());
+            QString decoded_user_id = UrlDecode(cookie.value().data());
+            migration_result_.mutableMiId() = decoded_user_id;
+        }
+        else if (cookie.name() == "migrated_user_id")
+        {
+            QString decoded_migrated_user_id = UrlDecode(cookie.value().data());
+            migration_result_.mutableDuokanId() = decoded_migrated_user_id;
         }
     }
     return true;
@@ -178,16 +209,33 @@ void XiaomiMigrationManager::onLoadFinished(bool ok)
     // TODO.
 }
 
+void XiaomiMigrationManager::onXiaomiExchangeFinished(bool ok)
+{
+    migration_result_.mutableToken() = account_manager_->xiaomiToken().token();
+    migration_result_.mutableMiId()  = account_manager_->xiaomiToken().userId();
+    if (ok)
+        emit migrationSucceeded();
+    else
+        emit migrationFailed();
+}
+
+void XiaomiMigrationManager::exchangeXiaomiToken()
+{
+    QUrl dushu_url = guessUrlFromString(MI_DUSHU_URI);
+    account_manager_->exchangeDuokanToken(dushu_url);
+}
 
 void XiaomiMigrationManager::onUrlChanged(const QUrl& url)
 {
     QString my_url = url.toEncoded();
+    qDebug("\nUrl Changed:%s\n", qPrintable(my_url));
     if (my_url.contains(config_.migrationServerOK()))
     {
+        view_->stop();
         loadDataFromCookies(url);
+        exchangeXiaomiToken();
         migration_result_.mutableMessage() = MIGRATION_MSG_SUCCEED;
         emit pageChanged(tr("Migration Succeeded"));
-        emit migrationSucceeded();
     }
     else if (my_url.contains(config_.migrationServerFail()))
     {
@@ -231,7 +279,7 @@ void XiaomiMigrationManager::onMigrationFailed()
 void XiaomiMigrationManager::saveResult()
 {
     migration_result_.saveToFile();
-    if (migration_result_.message() == MIGRATION_MSG_SUCCEED)
+    /*if (migration_result_.message() == MIGRATION_MSG_SUCCEED)
     {
         XiaomiToken mi_token;
         mi_token.mutableToken() = migration_result_.token();
@@ -239,7 +287,7 @@ void XiaomiMigrationManager::saveResult()
         mi_token.mutableUserId() = migration_result_.miId();
         mi_token.mutableCode() = QString("%1").arg(0);
         mi_token.saveToFile();
-    }
+    }*/
 }
 
 }
